+++ src/test/java/org/scion/jpan/demo/PingPongChannelServer.java
@@ -37,7 +37,7 @@ public class PingPongChannelServer {
     }
   }
 
-  public static void main(String[] args) throws UnknownHostException {
+  public static void main(String[] args) {
     try {
       service();
     } catch (SocketException ex) {