+++ src/test/java/org/scion/jpan/demo/ScmpTracerouteDemo.java
@@ -83,10 +83,7 @@ public class ScmpTracerouteDemo {
         }
       case PRODUCTION:
         {
-          //          System.setProperty(Constants.PROPERTY_DNS_SEARCH_DOMAINS, "ethz.ch.");
-
-          // Local port must be 30041 for networks that expect a dispatcher
-          ScmpTracerouteDemo demo = new ScmpTracerouteDemo(Constants.SCMP_PORT);
+          ScmpTracerouteDemo demo = new ScmpTracerouteDemo();
           demo.runDemo(ScionUtil.parseIA("64-2:0:44")); // VEX
           // demo.runDemo(ScionUtil.parseIA("66-2:0:10")); //singapore
           // demo.runDemo(DemoConstants.iaAnapayaHK);